# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Discutex API",
        "description": "API for managing threads, responses, and votes",
        "termsOfService": "https://www.example.com/terms",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/votes/{responseId}": {
            "post": {
                "tags": [
                    "Votes"
                ],
                "summary": "Register a vote on a response",
                "description": "Allows an authenticated user to vote on a response. Users cannot vote on their own responses.",
                "operationId": "vote",
                "parameters": [
                    {
                        "name": "responseId",
                        "in": "path",
                        "description": "ID of the response to vote on",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "action"
                                ],
                                "properties": {
                                    "action": {
                                        "description": "The vote action (true for upvote, false for downvote)",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vote registered successfully."
                    },
                    "400": {
                        "description": "Invalid response ID or invalid request data."
                    },
                    "403": {
                        "description": "User cannot vote on their own response."
                    },
                    "404": {
                        "description": "Response not found."
                    }
                }
            }
        },
        "/votes/unread": {
            "get": {
                "tags": [
                    "Votes"
                ],
                "summary": "Get all unread votes for the authenticated user",
                "description": "Retrieves all unread votes for the authenticated user, including the associated response details.",
                "operationId": "getUnreadVotes",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved unread votes.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "unreadVotes": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "user_id": {
                                                        "type": "integer"
                                                    },
                                                    "response_id": {
                                                        "type": "integer"
                                                    },
                                                    "type": {
                                                        "description": "The vote action (true for upvote, false for downvote)",
                                                        "type": "boolean"
                                                    },
                                                    "read": {
                                                        "description": "Whether the vote has been read by the user.",
                                                        "type": "boolean"
                                                    },
                                                    "created_at": {
                                                        "type": "string"
                                                    },
                                                    "updated_at": {
                                                        "type": "string"
                                                    },
                                                    "response": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "content": {
                                                                "type": "string"
                                                            },
                                                            "user_id": {
                                                                "type": "integer"
                                                            },
                                                            "thread_id": {
                                                                "type": "integer"
                                                            },
                                                            "created_at": {
                                                                "type": "string"
                                                            },
                                                            "updated_at": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized (User not authenticated)."
                    }
                }
            }
        }
    },
...
}